fix: consolidate openapitests package files under pkg-file-count limit
